pyproject.toml
[project]
name = "logfire-cli"
version = "0.1.0"
description = "CLI for interacting with Pydantic Logfire dashboards and content management"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Monitoring",
]
keywords = ["logfire", "pydantic", "dashboards", "perses", "observability", "monitoring"]
authors = [{ name = "Logfire Content Authors" }]

dependencies = [
    "aiohttp>=3.11",
    "pyyaml>=6.0",
    "pydantic>=2.10",
    "click>=8.1",
    "rich-click>=1.9",
    "rich>=13.9",
    "dirty-equals>=0.11",
]

[project.scripts]
logfire-cli = "logfire_cli.cli:cli"

[project.urls]
Homepage = "https://github.com/strawgate/logfire-content"
Repository = "https://github.com/strawgate/logfire-content"
Documentation = "https://strawgate.github.io/logfire-content/"

[build-system]
requires = ["uv_build>=0.9.25,<1.0.0"]
build-backend = "uv_build"

# ============================================================================
# UV Configuration
# ============================================================================

[tool.uv]
package = true

[tool.uv.build-backend]
module-root = "src"

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
    "basedpyright>=1.24",
    "aioresponses>=0.7",
    "inline-snapshot>=0.19",
    "dirty-equals>=0.8",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.27",
    "pymdown-extensions>=10.0",
    "mkdocs-click>=0.8",
]

# ============================================================================
# Ruff Configuration
# ============================================================================

[tool.ruff]
line-length = 140
target-version = "py312"
extend-exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
extend-select = [
    "A",      # flake8-builtins
    "ARG",    # flake8-unused-arguments
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "COM",    # flake8-commas
    "DTZ",    # flake8-datetimez
    "E",      # pycodestyle errors
    "EM",     # flake8-errmsg
    "F",      # pyflakes
    "FURB",   # refurb
    "I",      # isort
    "LOG",    # flake8-logging
    "N",      # pep8-naming
    "PERF",   # perflint
    "PIE",    # flake8-pie
    "PLR",    # pylint refactor
    "PLW",    # pylint warning
    "PT",     # flake8-pytest-style
    "PTH",    # flake8-use-pathlib
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "RSE",    # flake8-raise
    "RUF",    # ruff-specific
    "S",      # flake8-bandit (security)
    "SIM",    # flake8-simplify
    "TC",     # flake8-type-checking
    "TID",    # flake8-tidy-imports
    "TRY",    # tryceratops
    "UP",     # pyupgrade
    "W",      # pycodestyle warnings
    "D",      # pydocstyle
    "ANN",    # flake8-annotations
    "PLE",    # pylint errors
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "FA",     # flake8-future-annotations
    "PGH",    # pygrep-hooks
]
ignore = [
    "COM812",  # Conflicts with formatter (missing trailing comma)
    "D203",    # Conflicts with D211 (blank line before class docstring)
    "D213",    # Conflicts with D212 (multi-line summary second line)
    "D413",    # Missing blank line after last section
    "D100",    # Missing docstring in public module
    "A002",    # Argument shadowing builtin (sometimes unavoidable)
    "ANN401",  # Dynamically typed expressions (Any) - handled by Pyright
    "ISC001",  # Conflicts with formatter
    "Q000",    # Conflicts with formatter (single vs double quotes)
    "Q003",    # Conflicts with formatter (avoidable escaped quote)
]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["logfire_cli"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "S105",    # Hardcoded password strings OK in tests
    "S106",    # Hardcoded password arguments OK in tests
    "D",       # Docstrings not required in tests
    "ANN",     # Annotations not required in tests
    "ARG001",  # Unused arguments OK in tests (fixtures)
    "PLR2004", # Magic values OK in tests
]

[tool.ruff.lint.pydocstyle]
convention = "google"

# ============================================================================
# Pyright/Basedpyright Configuration
# ============================================================================

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "recommended"
reportImportCycles = "error"
reportMissingTypeStubs = false
reportExplicitAny = false

# ============================================================================
# Pytest Configuration
# ============================================================================

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-vv --inline-snapshot=create,fix"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (requires real API credentials)",
]

# ============================================================================
# Coverage Configuration
# ============================================================================

[tool.coverage.run]
source = ["src/logfire_cli"]
branch = true

[tool.coverage.report]
precision = 2
skip_covered = false
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
